# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tiketin Aja API",
        "description": "API Dokumentasi untuk Tiketin Aja (VanillaJS FE + Laravel 10 BE)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/allUser": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Ambil semua user",
                "description": "Mengambil semua data user",
                "operationId": "8281af6c1531bc7e5647362ed684695a",
                "responses": {
                    "200": {
                        "description": "List semua user"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Ambil user berdasarkan ID",
                "operationId": "568abef9fc85e19fbd30b5759bda7ebd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data user"
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    }
                }
            }
        },
        "/api/user/addUser": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Tambah user baru",
                "operationId": "9a30a58530a2962435f57fe566689303",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "age"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "age": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User berhasil ditambah"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/user/editUser/{id}": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Edit data user",
                "operationId": "c9ef01e042889f51c70f7cf42a4c3cd6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "age": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User berhasil diupdate"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/user/delUser/{id}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Hapus user berdasarkan ID",
                "operationId": "654c2b51a0f41cd625367375a145ac22",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User berhasil dihapus"
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "User",
            "description": "User"
        }
    ]
}